Fix sign-compare warnings
This fixes the following warning:

warning: comparison between signed and unsigned integer expressions
[-Wsign-compare]

Tested on Linux with the following command line:

$ bazel test :all

Reviewed-on: https://code-review.googlesource.com/3781

+++ re2/testing/exhaustive2_test.cc
@@ -23,7 +23,7 @@ TEST(EmptyString, Exhaustive) {
 TEST(Punctuation, Literals) {
   vector<string> alphabet = Explode("()*+?{}[]\\^$.");
   vector<string> escaped = alphabet;
-  for (int i = 0; i < escaped.size(); i++)
+  for (size_t i = 0; i < escaped.size(); i++)
     escaped[i] = "\\" + escaped[i];
   ExhaustiveTest(1, 1, escaped, RegexpGenerator::EgrepOps(),
                  2, alphabet, "", "");
